Bugfix/duplicate cropper

+++ Cropper/src/index.js
@@ -2,19 +2,6 @@ import _widgetBase from 'MxWidgetBase';
 import declare from 'dojoBaseDeclare';
 import * as widgetConf from '../conf/widget.config.json';
 import {
-    set as domAttrSet
-} from 'dojo/dom-attr';
-
-import {
-    create as domCreate,
-    place as domPlace,
-    empty as domEmpty
-} from 'dojo/dom-construct';
-
-import {
-    hitch
-} from 'dojo/_base/lang';
-import {
     isMxImageObject
 } from "./helpers/utils";
 
@@ -23,8 +10,8 @@ import $ from "jquery";
 require('imports-loader?jQuery=jquery!../node_modules/jquery-jcrop/js/jquery.Jcrop');
 require('imports-loader?jQuery=jquery!../node_modules/jquery-jcrop/js/jquery.color');
 import './style/style.scss';
-import "../node_modules/jquery-jcrop/css/Jcrop.gif";
-import "../node_modules/jquery-jcrop/css/jquery.Jcrop.min.css";
+import "jquery-jcrop/css/Jcrop.gif";
+import "jquery-jcrop/css/jquery.Jcrop.min.css";
 
 
 export default declare(`${widgetConf.name}.widget.${widgetConf.name}`, [_widgetBase], {
@@ -38,85 +25,81 @@ export default declare(`${widgetConf.name}.widget.${widgetConf.name}`, [_widgetB
     JCropAPI: null,
 
     postCreate() {
-        console.debug(`${this.id} >> postCreate`);
+        logger.debug(`${this.id} >> postCreate`);
         $(this.domNode).addClass("jcrop-mx-widget");
-        domAttrSet(this.domNode, "tabIndex", -1);
+        $(this.domNode).attr("tabIndex", -1);
+
+        this.JCropAPI = null;
+        this.addOnDestroy(() => {
+            this.JCropAPI && this.JCropAPI.destroy && this.JCropAPI.destroy();
+        });
     },
 
     update(contextObject, callback) {
-        console.debug(`${this.id} >> update`);
+        logger.debug(`${this.id} >> update`);
 
         if (contextObject && isMxImageObject(contextObject)) {
             this._contextObject = contextObject;
             this.fileID = this._contextObject.get('FileID');
-            this._updateRendering();
+            this._updateRendering(callback);
             this._addSubscriptions();
         } else {
             this._handleError(`${widgetConf.name} should be initiated in a nonempty context object that inherits from 'System.Image' Entity.`);
-        }
-        if (callback && typeof callback === "function") {
-            callback();
+            this._executeCallback(callback);
         }
     },
 
     _addSubscriptions() {
-        console.debug(`${this.id} >> _addSubscriptions`);
+        logger.debug(`${this.id} >> _addSubscriptions`);
         this.unsubscribeAll();
         this.subscribe({
             guid: this._contextObject.getGuid(),
-            callback: hitch(this, () => {
-                console.debug(`${this.id} >> subscription has been set successfully`);
+            callback: $.proxy(() => {
+                logger.debug(`${this.id} >> subscription has been set successfully`);
                 this._updateRendering();
-            })
+            }, this)
         });
     },
 
-    _updateRendering() {
-        console.debug(`${this.id} >> _updateRendering`);
+    _updateRendering(callback) {
+        logger.debug(`${this.id} >> _updateRendering`);
         var src = '/file?fileID=' + this.fileID + '&' + (+new Date()).toString(36);
         if (this.imgNode === null) {
-            domEmpty(this.domNode);
-            this.imgNode = domCreate('img', {
-                src
-            });
-            domPlace(this.imgNode, this.domNode);
+            $(this.domNode).empty();
+            this.imgNode = $("<img>").attr("src", src).css("display", "none");
+            $(this.imgNode).appendTo(this.domNode);
         } else {
-            domAttrSet(this.imgNode, 'src', src);
+            $(this.imgNode).attr("src", src);
         }
-        if (this.JCropAPI) {
-            // destroy current JCrop instance, in order to reset 
+        this._initJCrop(callback);
+    },
+
+    _initJCrop(callback) {
+        logger.debug(`${this.id} >> _initJCrop`);
+        if (this.JCropAPI !== null && this.JCropAPI.destroy) {
+            // destroy current JCrop instance, in order to reset
             this.JCropAPI.destroy();
+            this.JCropAPI = null;
         }
-        this._initJCrop();
-    },
 
-    _initJCrop() {
-        console.debug(`${this.id} >> _initJCrop`);
-        var widgetSelfRef = this;
+        var widget = this;
         var cropOptions = this._getCroppingOptions();
-        $(this.imgNode).Jcrop(cropOptions, function () {
-            console.debug(`${widgetSelfRef.id} >> _getReferenceToJCropInstance`);
-            widgetSelfRef.JCropAPI = this;
-        });
-    },
 
-    uninitialize() {
-        logger.debug('cropper.widget.cropper uninitialize');
-        this.unsubscribeAll();
+        $(this.imgNode).Jcrop(cropOptions, function () {
+            logger.debug(`${widget.id} >> _getReferenceToJCropInstance`);
+            widget.JCropAPI = this;
+            widget._executeCallback(callback);
+        });
     },
 
     _handleError(errorMessage) {
-        console.debug(`${this.id} >> _handleError`);
-        domEmpty(this.domNode);
-        const errorMessageNode = domCreate("div", {
-            class: "alert alert-danger",
-            innerText: errorMessage
-        });
-        domPlace(errorMessageNode, this.domNode);
+        logger.debug(`${this.id} >> _handleError`);
+        $(this.domNode).empty();
+        $("<div>").addClass("alert alert-danger").text(errorMessage).appendTo(this.domNode);
     },
 
     _setCroppingCoordinates(coordinates) {
-        console.debug(`${this.id} >> _setCroppingCoordinates`);
+        logger.debug(`${this.id} >> _setCroppingCoordinates`);
         if (coordinates) {
             this._contextObject.set('crop_x1', Math.round(coordinates.x));
             this._contextObject.set('crop_x2', Math.round(coordinates.x2));
@@ -128,14 +111,14 @@ export default declare(`${widgetConf.name}.widget.${widgetConf.name}`, [_widgetB
 
     },
     _getCroppingOptions() {
-        console.debug(`${this.id} >> _getCroppingOptions`);
+        logger.debug(`${this.id} >> _getCroppingOptions`);
         var options = {};
         options.aspectRatio = this._getAspectRatio();
         options.bgColor = 'black';
         options.bgOpacity = 0.4;
-        options.onSelect = hitch(this, this._setCroppingCoordinates);
-        options.onChange = hitch(this, this._setCroppingCoordinates);
-        options.onRelease = hitch(this, this._setCroppingCoordinates);
+        options.onSelect = $.proxy(this._setCroppingCoordinates, this);
+        options.onChange = $.proxy(this._setCroppingCoordinates, this);
+        options.onRelease = $.proxy(this._setCroppingCoordinates, this);
         options.setSelect = [0, 0, this.startwidth, this.startheight];
         options.boxWidth = this.cropwidth;
         options.boxHeight = this.cropheight;
@@ -143,7 +126,7 @@ export default declare(`${widgetConf.name}.widget.${widgetConf.name}`, [_widgetB
         return options;
     },
     _getAspectRatio() {
-        console.debug(`${this.id} >> _getAspectRatio`);
+        logger.debug(`${this.id} >> _getAspectRatio`);
         var aspectRatio = null,
             aspectRatioArr = null,
             givenWidth = null,
@@ -153,6 +136,12 @@ export default declare(`${widgetConf.name}.widget.${widgetConf.name}`, [_widgetB
         givenHeight = parseInt(aspectRatioArr[1], 10);
         aspectRatio = givenWidth / givenHeight;
         return aspectRatio;
+    },
+
+    _executeCallback(callback) {
+        if (callback && typeof callback === "function") {
+            callback();
+        }
     }
 
 });
